+++ src/components/builds/weapons/WeaponCard.tsx
@@ -6,7 +6,7 @@ import { IWeaponData, IWeaponDb } from '../../../data/types';
 function WeaponCard({ oid, name, rarity, popularity }: IWeaponData & { popularity: number }) {
   return (
     <div className={`weapon-card`}>
-      <img className={`rarity-${rarity}`} src={`/assets/weapons/${oid}.png`} alt={name} />
+      <img className={`rarity-${rarity}`} src={`/assets/weapons/${oid}.webp`} alt={name} />
       <div className="weapon-detail">
         <div className="weapon-name">
           {name}